rename pokt to cosmos
